Refactor account directory structure
