+++ web/src/App.tsx
@@ -1,15 +1,10 @@
 import React from 'react';
 import 'bootstrap/dist/css/bootstrap.css';
-import {BrowserRouter, Navigate, Route, Routes} from 'react-router-dom';
+import {BrowserRouter, Route, Routes} from 'react-router-dom';
 import {Home} from './components/home/home';
-import {Profile} from './components/profile';
-import {Security} from './components/security';
 import {Error} from './components/shared/error';
-import {Loading} from './components/shared/loading';
 import {NotFound} from './components/shared/notFound';
-import {RequireAuth} from "./components/auth/requireAuth";
 import {Oidc} from "./components/auth/oidc";
-import {Layout} from "./components/layout";
 import {Register} from "./components/register/register";
 import {ForgotPassword} from "./components/forgotPassword/forgotPassword";
 import {ResetPassword} from "./components/resetPassword/resetPassword";
@@ -17,6 +12,7 @@ import {ConfirmEmail} from "./components/confirmEmail/confirmEmail";
 import {Login} from "./components/login/login";
 import {Consent} from "./components/consent/consent";
 import {Logout} from "./components/logout/logout";
+import {Account} from "./components/account";
 
 function App() {
     return (
@@ -35,26 +31,9 @@ function App() {
                 <Route path="/error" element={<Error/>}/>
                 <Route path="/404" element={<NotFound/>}/>
                 <Route path="*" element={<NotFound/>}/>
-                {/* For Debug */}
-                <Route path="/loading" element={<Loading/>}/>
             </Routes>
         </BrowserRouter>
-    )
-}
-
-function Account() {
-    return (
-        <RequireAuth>
-            <Layout>
-                <Routes>
-                    <Route index element={<Navigate to={"/account/profile"}/>}/>
-                    <Route path="profile" element={<Profile/>}/>
-                    <Route path="security/*" element={<Security/>}/>
-                    <Route path="*" element={<Navigate to={"/404"}/>}/>
-                </Routes>
-            </Layout>
-        </RequireAuth>
-    )
+    );
 }
 
 export default App;
