+++ web/src/components/account/profile/info.tsx
@@ -2,7 +2,7 @@ import {Alert, Button, Form, Modal} from "react-bootstrap";
 import React from "react";
 import {useTranslation} from "react-i18next";
 import {User} from "oidc-client-ts";
-import config from "../../config.json";
+import config from "../../../config/config";
 
 export function Info(props: {
     show: boolean,
@@ -63,7 +63,7 @@ export function Info(props: {
 }
 
 function putNameUpdate(name: string, user: User) {
-    return fetch(`${config.server_url}/api/account/info`, {
+    return fetch(`${config.api_url}/account/info`, {
         method: "PUT",
         headers: {
             "Content-Type": "application/json",
