+++ web/src/components/account/profile/phone.tsx
@@ -1,10 +1,10 @@
 import React, {useEffect} from "react";
 import {useTranslation} from "react-i18next";
 import {Alert, Button, Collapse, Form, InputGroup, Modal} from "react-bootstrap";
-import {phonePatternTest} from "../../utils/regex";
+import {phonePatternTest} from "../../../utils/regex";
 import {User} from "oidc-client-ts";
-import config from "../../config.json";
-import i18n from "../../i18n/i18n";
+import config from "../../../config/config";
+import i18n from "../../../i18n/i18n";
 
 export function Phone(props: {
     show: boolean,
@@ -173,7 +173,7 @@ export function Phone(props: {
 }
 
 function postSendVerificationCode(phone: string, user: User) {
-    return fetch(`${config.server_url}/api/account/phone`, {
+    return fetch(`${config.api_url}/account/phone`, {
         method: "POST",
         headers: {
             "Content-Type": "application/json",
@@ -193,7 +193,7 @@ function postSendVerificationCode(phone: string, user: User) {
 }
 
 function putPhoneUpdate(phone: string, code: string, user: User) {
-    return fetch(`${config.server_url}/api/account/phone`, {
+    return fetch(`${config.api_url}/account/phone`, {
         method: "PUT",
         headers: {
             "Content-Type": "application/json",
