+++ web/src/components/account/profile/picture.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import {useTranslation} from "react-i18next";
 import {Alert, Button, Form, Image, Modal} from "react-bootstrap";
-import userIcon from "../../img/user.svg";
+import userIcon from "../../../img/user.svg";
 import {User} from "oidc-client-ts";
 
 export function Picture(props: { show: boolean, user: User | null, handleHide: () => void }) {
