+++ web/src/config/config.json
@@ -1,0 +1,10 @@
+{
+    "public_url": "http://127.0.0.1",
+    "default_language": "zh-cn",
+    "api_url": "http://127.0.0.1/api",
+    "oidc": {
+        "authority": "http://127.0.0.1",
+        "client_secret": "client_secret",
+        "redirect_uri": "http://127.0.0.1/oidc"
+    }
+}
